ml_models/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mlflow==2.10.2
psycopg2-binary==2.9.9
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
torch==2.2.0
arch==6.3.0
pydantic==2.5.3
